requirements.txt
hikari
hikari-tanjun
